# usoft-backend/src/docs/openapi-usof.yaml
openapi: 3.0.3
info:
  title: USOF Backend API
  version: 1.2.0
  description: |
    Forum platform backend (No ORM, pure SQL + mysql2). This spec enumerates all implemented routes, parameters, request/response schemas, and auth rules.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    AuthToken:
      type: object
      properties:
        token:
          type: string
      required: [token]
    Ok:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Category:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        slug: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, createdAt, updatedAt]
    CategoryCreate:
      type: object
      properties:
        title: { type: string, minLength: 2 }
        description: { type: string, nullable: true }
      required: [title]
    CategoryUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
    Post:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        status: { type: string, enum: [active, inactive] }
        likesCount: { type: integer }
        dislikesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        authorId: { type: integer }
        authorLogin: { type: string }
        authorFullName: { type: string }
        authorRating: { type: integer }
      required: [id, title, content, status, likesCount, dislikesCount, createdAt, updatedAt]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user (sends email confirmation token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password, passwordConfirm, fullName, email]
              properties:
                login: { type: string }
                password: { type: string }
                passwordConfirm: { type: string }
                fullName: { type: string }
                email: { type: string, format: email }
      responses:
        '201':
          description: Created (user id + echo)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  login: { type: string }
                  email: { type: string, format: email }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and get JWT (requires login, email, password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, email, password]
              properties:
                login: { type: string }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: JWT token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthToken' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/auth/confirm/{token}:
    get:
      tags: [Auth]
      summary: Confirm email by token
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Email confirmed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/auth/password-reset:
    post:
      tags: [Auth]
      summary: Start password reset (sends email with token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
              required: [email]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
  /api/auth/password-reset/{token}:
    post:
      tags: [Auth]
      summary: Confirm password reset by token
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, minLength: 6 }
              required: [password]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }

  # ===================== Users ======================
...
  /api/posts/{id}/categories:
    get:
      tags: [Posts, Categories]
      summary: List categories for a post
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
...
  /api/categories:
    get:
      tags: [Categories]
      summary: List categories
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }

    post:
      tags: [Categories]
      summary: Create category (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryCreate' }
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/categories/{id}:
    get:
      tags: [Categories]
      summary: Get category by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Categories]
      summary: Update category (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryUpdate' }
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Categories]
      summary: Delete category (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/categories/{id}/posts:
    get:
      tags: [Categories, Posts]
      summary: List posts belonging to a category
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Posts in category
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Post' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===================== Admin (REST) ===============
...
  /api/admin/categories:
    get:
      tags: [Admin]
      summary: List categories
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
    post:
      tags: [Admin]
      summary: Create category
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string, nullable: true }
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
  /api/admin/categories/{id}:
    patch:
      tags: [Admin]
      summary: Update category (title/description; slug recalculated when title provided)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string, nullable: true }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Admin]
      summary: Delete category
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
...